separate layers and add MLP
