+++ src/algos/sac.py
@@ -10,6 +10,7 @@ from torch_geometric.nn import GCNConv
 from torch_geometric.utils import grid
 from src.algos.reb_flow_solver import solveRebFlow
 from src.misc.utils import dictsum
+from src.algos.layers import GNNActor, GNNActor1, MLPActor, GNNCritic1, GNNCritic2, GNNCritic3, GNNCritic4, GNNCritic5, GNNCritic6, MLPCritic4
 import random
 import json
 
@@ -89,295 +90,6 @@ class Scalar(nn.Module):
 
     def forward(self):
         return self.constant
-
-
-#########################################
-############## ACTOR ####################
-#########################################
-class GNNActor(nn.Module):
-    """
-    Actor \pi(a_t | s_t) parametrizing the concentration parameters of a Dirichlet Policy.
-    """
-
-    def __init__(self, in_channels, hidden_size=32, act_dim=6, mode=0):
-        super().__init__()
-        self.mode = mode
-        self.in_channels = in_channels
-        self.act_dim = act_dim
-        self.conv1 = GCNConv(in_channels, in_channels)
-        self.lin1 = nn.Linear(in_channels, hidden_size)
-        self.lin2 = nn.Linear(hidden_size, hidden_size)
-        if mode == 0:
-            self.lin3 = nn.Linear(hidden_size, 1)
-        elif mode == 1:
-            self.lin3 = nn.Linear(hidden_size, 2)
-        else:
-            self.lin3 = nn.Linear(hidden_size, 3)
-
-    def forward(self, state, edge_index, deterministic=False):
-        out = F.relu(self.conv1(state, edge_index))
-        x = out + state
-        x = x.reshape(-1, self.act_dim, self.in_channels)
-        x = F.leaky_relu(self.lin1(x))
-        x = F.leaky_relu(self.lin2(x))
-        x = F.softplus(self.lin3(x))
-        concentration = x.squeeze(-1)
-        if deterministic:
-            if self.mode == 0:
-                action = (concentration) / (concentration.sum() + 1e-20)
-            elif self.mode == 1:
-                # action_o = (concentration[:,:,0]-1)/(concentration[:,:,0] + concentration[:,:,1] -2 + 1e-20)
-                # action_d = (concentration[:,:,2]-1)/(concentration[:,:,2] + concentration[:,:,3] -2 + 1e-20)
-                # action = torch.cat((action_o.squeeze(0).unsqueeze(-1), action_d.squeeze(0).unsqueeze(-1)),-1)
-                action_o = (concentration[:,:,0]-1)/(concentration[:,:,0] + concentration[:,:,1] -2 + 1e-20)
-                action = action_o.squeeze(0).unsqueeze(-1)
-            else:
-                # action_o = (concentration[:,:,0]-1)/(concentration[:,:,0] + concentration[:,:,1] -2 + 1e-20)
-                # action_d = (concentration[:,:,2]-1)/(concentration[:,:,2] + concentration[:,:,3] -2 + 1e-20)
-                # action_reb = (concentration[:,:,4]) / (concentration[:,:,4].sum() + 1e-20)
-                # action = torch.cat((action_o.squeeze(0).unsqueeze(-1), action_d.squeeze(0).unsqueeze(-1),action_reb.squeeze(0).unsqueeze(-1)),-1)
-                action_o = (concentration[:,:,0]-1)/(concentration[:,:,0] + concentration[:,:,1] -2 + 1e-20)
-                action_reb = (concentration[:,:,4]) / (concentration[:,:,4].sum() + 1e-20)
-                action = torch.cat((action_o.squeeze(0).unsqueeze(-1), action_reb.squeeze(0).unsqueeze(-1)),-1)
-            log_prob = None
-        else:
-            if self.mode == 0:
-                m = Dirichlet(concentration + 1e-20)
-                action = m.rsample()
-                log_prob = m.log_prob(action)
-                action = action.squeeze(0).unsqueeze(-1)
-            elif self.mode == 1:
-                # m_o = Beta(concentration[:,:,0] + 1e-20, concentration[:,:,1] + 1e-20)
-                # m_d = Beta(concentration[:,:,2] + 1e-20, concentration[:,:,3] + 1e-20)
-                # action_o = m_o.rsample()
-                # action_d = m_d.rsample()
-                # log_prob = m_o.log_prob(action_o).sum(dim=-1) + m_d.log_prob(action_d).sum(dim=-1)
-                # action = torch.cat((action_o.squeeze(0).unsqueeze(-1), action_d.squeeze(0).unsqueeze(-1)),-1)
-                m_o = Beta(concentration[:,:,0] + 1e-20, concentration[:,:,1] + 1e-20)
-                action_o = m_o.rsample()
-                log_prob = m_o.log_prob(action_o).sum(dim=-1)
-                action = action_o.squeeze(0).unsqueeze(-1)             
-            else:
-                # Price
-                # m_o = Beta(concentration[:,:,0] + 1e-20, concentration[:,:,1] + 1e-20)
-                # m_d = Beta(concentration[:,:,2] + 1e-20, concentration[:,:,3] + 1e-20)
-                # action_o = m_o.rsample()
-                # action_d = m_d.rsample()
-                # # Rebalancing desired distribution
-                # m_reb = Dirichlet(concentration[:,:,-1] + 1e-20)
-                # action_reb = m_reb.rsample()              
-                # log_prob = m_o.log_prob(action_o).sum(dim=-1) + m_d.log_prob(action_d).sum(dim=-1) + m_reb.log_prob(action_reb)
-                # action = torch.cat((action_o.squeeze(0).unsqueeze(-1), action_d.squeeze(0).unsqueeze(-1),action_reb.squeeze(0).unsqueeze(-1)),-1)         
-                m_o = Beta(concentration[:,:,0] + 1e-20, concentration[:,:,1] + 1e-20)
-                action_o = m_o.rsample()
-                # Rebalancing desired distribution
-                m_reb = Dirichlet(concentration[:,:,-1] + 1e-20)
-                action_reb = m_reb.rsample()              
-                log_prob = m_o.log_prob(action_o).sum(dim=-1) + m_reb.log_prob(action_reb)
-                action = torch.cat((action_o.squeeze(0).unsqueeze(-1), action_reb.squeeze(0).unsqueeze(-1)),-1)       
-        return action, log_prob
-    
-
-class MLPActor(nn.Module):
-    """
-    Actor \pi(a_t | s_t) parametrizing the concentration parameters of a Dirichlet Policy.
-    """
-
-    def __init__(self, in_channels, hidden_size=128, act_dim=10, mode=1):
-        super().__init__()
-        self.in_channels = in_channels
-        self.nregion = act_dim
-        self.conv1 = GCNConv(in_channels, in_channels)
-        self.lin1 = nn.Linear(in_channels, hidden_size)
-        self.lin2 = nn.Linear(hidden_size, 2*self.nregion)
-
-    def forward(self, state, edge_index, deterministic=False):
-        out = F.relu(self.conv1(state, edge_index))
-        x = out + state
-        x = x.reshape(-1, self.nregion, self.in_channels)
-        x = F.leaky_relu(self.lin1(x))
-        x = F.softplus(self.lin2(x))
-        concentration = x.squeeze(-1)
-        if deterministic:
-            action = (concentration[:,:,:self.nregion]-1)/(concentration[:,:,:self.nregion] + concentration[:,:,self.nregion:] -2 + 1e-20)
-        else:
-            # m_o = Beta(concentration[:,:,0] + 1e-20, concentration[:,:,1] + 1e-20)
-            # m_d = Beta(concentration[:,:,2] + 1e-20, concentration[:,:,3] + 1e-20)
-            # action_o = m_o.rsample()
-            # action_d = m_d.rsample()
-            # log_prob = m_o.log_prob(action_o).sum(dim=-1) + m_d.log_prob(action_d).sum(dim=-1)
-            # action = torch.cat((action_o.squeeze(0).unsqueeze(-1), action_d.squeeze(0).unsqueeze(-1)),-1)
-            m = Beta(concentration[:,:,:self.nregion] + 1e-20, concentration[:,:,self.nregion:] + 1e-20)
-            action = m.rsample().squeeze(0)
-            log_prob = m.log_prob(action).sum(dim=-1).sum(dim=-1)          
-        return action, log_prob
-
-#########################################
-############## CRITIC ###################
-#########################################
-
-
-class GNNCritic1(nn.Module):
-    """
-    Architecture 1, GNN, Pointwise Multiplication, Readout, FC
-    """
-
-    def __init__(self, in_channels, hidden_size=256, act_dim=6):
-        super().__init__()
-        self.act_dim = act_dim
-        self.in_channels = in_channels
-        self.conv1 = GCNConv(in_channels, in_channels)
-        self.lin1 = nn.Linear(in_channels, hidden_size)
-        self.lin2 = nn.Linear(hidden_size, hidden_size)
-        self.lin3 = nn.Linear(hidden_size, 1)
-
-    def forward(self, state, edge_index, action):
-        out = F.relu(self.conv1(state, edge_index))
-        x = out + state
-        x = x.reshape(-1, self.act_dim, self.in_channels)  # (B,N,21)
-        action = action * 10
-        action = action.unsqueeze(-1)  # (B,N,1)
-        x = x * action  # pointwise multiplication (B,N,21)
-        x = x.sum(dim=1)  # (B,21)
-        x = F.relu(self.lin1(x))  # (B,H)
-        x = F.relu(self.lin2(x))  # (B,H)
-        x = self.lin3(x).squeeze(-1)  # (B)
-        return x
-
-
-class GNNCritic2(nn.Module):
-    """
-    Architecture 2, GNN, Readout, Concatenation, FC
-    """
-
-    def __init__(self, in_channels, hidden_size=256, act_dim=6):
-        super().__init__()
-        self.act_dim = act_dim
-        self.conv1 = GCNConv(in_channels, in_channels)
-        self.lin1 = nn.Linear(in_channels + act_dim, hidden_size)
-        self.lin2 = nn.Linear(hidden_size, hidden_size)
-        self.lin3 = nn.Linear(hidden_size, 1)
-
-    def forward(self, state, edge_index, action):
-        out = F.relu(self.conv1(state, edge_index))
-        x = out + state
-        x = x.reshape(-1, self.act_dim, 21)  # (B,N,21)
-        x = torch.sum(x, dim=1)  # (B, 21)
-        concat = torch.cat([x, action], dim=-1)  # (B, 21+N)
-        x = F.relu(self.lin1(concat))  # (B,H)
-        x = F.relu(self.lin2(x))  # (B,H)
-        x = self.lin3(x).squeeze(-1)  # B
-        return x
-
-
-class GNNCritic3(nn.Module):
-    """
-    Architecture 3: Concatenation, GNN, Readout, FC
-    """
-
-    def __init__(self, in_channels, hidden_size=32, act_dim=6):
-        super().__init__()
-        self.act_dim = act_dim
-        self.conv1 = GCNConv(22, 22)
-        self.lin1 = nn.Linear(22, hidden_size)
-        self.lin2 = nn.Linear(hidden_size, hidden_size)
-        self.lin3 = nn.Linear(hidden_size, 1)
-
-    def forward(self, state, edge_index, action):
-        cat = torch.cat([state, action.unsqueeze(-1)], dim=-1)  # (B,N,22)
-        out = F.relu(self.conv1(cat, edge_index))
-        x = out + cat
-        x = x.reshape(-1, self.act_dim, 22)  # (B,N,22)
-        x = F.relu(self.lin1(x))  # (B, H)
-        x = F.relu(self.lin2(x))  # (B, H)
-        x = torch.sum(x, dim=1)  # (B, 22)
-        x = self.lin3(x).squeeze(-1)  # (B)
-        return x
-
-
-class GNNCritic4(nn.Module):
-    """
-    Architecture 4: GNN, Concatenation, FC, Readout
-    """
-
-    def __init__(self, in_channels, hidden_size=32, act_dim=6, mode=1):
-        super().__init__()
-        self.act_dim = act_dim
-        self.mode = mode
-        self.conv1 = GCNConv(in_channels, in_channels)
-        # self.lin1 = nn.Linear(in_channels + self.mode + 1, hidden_size)
-        if (mode == 0) | (mode == 1):
-            self.lin1 = nn.Linear(in_channels + 1, hidden_size)
-        else:
-            self.lin1 = nn.Linear(in_channels + 2, hidden_size)
-        self.lin2 = nn.Linear(hidden_size, hidden_size)
-        self.lin3 = nn.Linear(hidden_size, 1)
-        self.in_channels = in_channels
-
-    def forward(self, state, edge_index, action):
-        out = F.relu(self.conv1(state, edge_index))
-        x = out + state
-        x = x.reshape(-1, self.act_dim, self.in_channels)  # (B,N,21)
-        concat = torch.cat([x, action], dim=-1)  # (B,N,22)
-        x = F.relu(self.lin1(concat))
-        x = F.relu(self.lin2(x))  # (B, N, H)
-        x = torch.sum(x, dim=1)  # (B, H)
-        x = self.lin3(x).squeeze(-1)  # (B)
-        return x
-
-
-class GNNCritic5(nn.Module):
-    """
-    Architecture 5, GNN, Pointwise Multiplication, FC, Readout
-    """
-
-    def __init__(self, in_channels, hidden_size=256, act_dim=6):
-        super().__init__()
-        self.act_dim = act_dim
-        self.in_channels = in_channels
-        self.conv1 = GCNConv(in_channels, in_channels)
-        self.lin1 = nn.Linear(in_channels, hidden_size)
-        self.lin2 = nn.Linear(hidden_size, hidden_size)
-        self.lin3 = nn.Linear(hidden_size, 1)
-
-    def forward(self, state, edge_index, action):
-        out = F.relu(self.conv1(state, edge_index))
-        x = out + state
-        x = x.reshape(-1, self.act_dim, self.in_channels)  # (B,N,21)
-        action = action + 1
-        action = action.unsqueeze(-1)  # (B,N,1)
-        x = x * action  # pointwise multiplication (B,N,21)
-        x = F.relu(self.lin1(x))  # (B,N,H)
-        x = F.relu(self.lin2(x))  # (B,N,H)
-        x = x.sum(dim=1)  # (B,H)
-        x = self.lin3(x).squeeze(-1)  # (B)
-        return x
-    
-
-class MLPCritic4(nn.Module):
-    """
-    Architecture 4: FC, Concatenation, FC, Readout
-    """
-
-    def __init__(self, in_channels, hidden_size=128, act_dim=10, mode=1):
-        super().__init__()
-        self.nregion = act_dim
-        self.conv1 = GCNConv(in_channels, in_channels)
-        self.lin1 = nn.Linear(in_channels + self.nregion, hidden_size)
-        self.lin2 = nn.Linear(hidden_size, hidden_size)
-        self.lin3 = nn.Linear(hidden_size, 1)
-        self.in_channels = in_channels
-
-    def forward(self, state, edge_index, action):
-        out = F.relu(self.conv1(state, edge_index))
-        x = out + state
-        x = x.reshape(-1, self.nregion, self.in_channels)  # (B,N,21)
-        concat = torch.cat([x, action], dim=-1)  # (B,N,22)
-        x = F.relu(self.lin1(concat))
-        x = F.relu(self.lin2(x))  # (B, N, H)
-        x = torch.sum(x, dim=1)  # (B, H)
-        x = self.lin3(x).squeeze(-1)  # (B)
-        return x
 
 
 #########################################
